Add files via upload
Let's take an example of two sentences, "I have a stick in my hand" and "Let's stick to the plan". I, a human, can very well understand that the word "Stick" means two different things depending on the context that we consider. But the question is, do we have a neural network that understands this as well? Yes we do and we call it BERT!!